+++ ta4j-core/src/main/java/org/ta4j/core/criteria/commissions/CommissionsCriterion.java
@@ -1,4 +1,27 @@
-package org.ta4j.core.criteria.commission;
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2017-2025 Ta4j Organization & respective
+ * authors (see AUTHORS)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package org.ta4j.core.criteria.commissions;
 
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.Position;
@@ -9,17 +32,19 @@ import org.ta4j.core.num.Num;
 /**
  * Analysis criterion that sums all commissions paid across positions.
  *
- * <p>The criterion relies on each position cost model to determine the paid
- * commission and adds them together for a trading record.</p>
+ * <p>
+ * The criterion relies on each position cost model to determine the paid
+ * commission and adds them together for a trading record.
+ * </p>
  *
  * @since 0.19
  */
-public class CommissionCriterion extends AbstractAnalysisCriterion {
+public class CommissionsCriterion extends AbstractAnalysisCriterion {
 
     /**
      * Calculates the commission paid for a single position.
      *
-     * @param series the bar series used for number creation
+     * @param series   the bar series used for number creation
      * @param position the evaluated position
      * @return the commission paid for the provided position or zero when it is open
      * @since 0.19
