+++ ta4j-core/src/main/java/org/ta4j/core/criteria/commissions/CommissionsImpactPercentageCriterion.java
@@ -1,4 +1,27 @@
-package org.ta4j.core.criteria.commission;
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2017-2025 Ta4j Organization & respective
+ * authors (see AUTHORS)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package org.ta4j.core.criteria.commissions;
 
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.Position;
@@ -10,15 +33,17 @@ import org.ta4j.core.num.Num;
 /**
  * Analysis criterion that measures how much commission reduces gross profit.
  *
- * <p>It expresses the commission impact as the percentage of the gross profit or
- * loss.</p>
+ * <p>
+ * It expresses the commission impact as the percentage of the gross profit or
+ * loss.
+ * </p>
  *
  * @since 0.19
  */
-public final class CommissionImpactPercentageCriterion extends AbstractAnalysisCriterion {
+public final class CommissionsImpactPercentageCriterion extends AbstractAnalysisCriterion {
 
     private static final GrossProfitLossCriterion GROSS_PROFIT_LOSS_CRITERION = new GrossProfitLossCriterion();
-    private static final CommissionCriterion COMMISSION_CRITERION = new CommissionCriterion();
+    private static final CommissionsCriterion COMMISSION_CRITERION = new CommissionsCriterion();
 
     /**
      * Calculates the commission percentage impact for a single position.
