+++ ta4j-core/src/main/java/org/ta4j/core/criteria/pnl/MaxPositionNetLossCriterion.java
@@ -1,3 +1,26 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2017-2025 Ta4j Organization & respective
+ * authors (see AUTHORS)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package org.ta4j.core.criteria.pnl;
 
 import org.ta4j.core.BarSeries;
@@ -7,11 +30,12 @@ import org.ta4j.core.criteria.AbstractAnalysisCriterion;
 import org.ta4j.core.num.Num;
 
 /**
- * Analysis criterion that identifies the worst net loss among closed
- * positions.
+ * Analysis criterion that identifies the worst net loss among closed positions.
  *
- * <p>The criterion returns the most negative profit value found within a
- * trading record.</p>
+ * <p>
+ * The criterion returns the most negative profit value found within a trading
+ * record.
+ * </p>
  *
  * @since 0.19
  */
@@ -21,7 +45,7 @@ public final class MaxPositionNetLossCriterion extends AbstractAnalysisCriterion
      * Returns the net profit of the provided position.
      *
      * @param barSeries the bar series used for number creation
-     * @param position the evaluated position
+     * @param position  the evaluated position
      * @return the net profit, which will be negative for a loss
      * @since 0.19
      */
@@ -31,10 +55,9 @@ public final class MaxPositionNetLossCriterion extends AbstractAnalysisCriterion
     }
 
     /**
-     * Finds the largest net loss among all closed positions in the trading
-     * record.
+     * Finds the largest net loss among all closed positions in the trading record.
      *
-     * @param barSeries the bar series used for number creation
+     * @param barSeries     the bar series used for number creation
      * @param tradingRecord the trading record containing the positions to scan
      * @return the most negative profit value or zero when every position is
      *         profitable or break-even
@@ -42,7 +65,8 @@ public final class MaxPositionNetLossCriterion extends AbstractAnalysisCriterion
      */
     @Override
     public Num calculate(BarSeries barSeries, TradingRecord tradingRecord) {
-        return tradingRecord.getPositions().stream()
+        return tradingRecord.getPositions()
+                .stream()
                 .filter(Position::isClosed)
                 .map(Position::getProfit)
                 .filter(Num::isNegative)
