+++ ta4j-core/src/main/java/org/ta4j/core/criteria/pnl/MaxPositionNetProfitCriterion.java
@@ -1,3 +1,26 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2017-2025 Ta4j Organization & respective
+ * authors (see AUTHORS)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package org.ta4j.core.criteria.pnl;
 
 import org.ta4j.core.BarSeries;
@@ -10,8 +33,9 @@ import org.ta4j.core.num.Num;
  * Analysis criterion that identifies the best net profit among closed
  * positions.
  *
- * <p>The criterion returns the highest profit value found within a trading
- * record.</p>
+ * <p>
+ * The criterion returns the highest profit value found within a trading record.
+ * </p>
  *
  * @since 0.19
  */
@@ -34,14 +58,15 @@ public final class MaxPositionNetProfitCriterion extends AbstractAnalysisCriteri
      * Finds the largest net profit among all closed positions in the trading
      * record.
      *
-     * @param barSeries the bar series used for number creation
+     * @param barSeries     the bar series used for number creation
      * @param tradingRecord the trading record containing the positions to scan
      * @return the highest profit value or zero when no position ends in profit
      * @since 0.19
      */
     @Override
     public Num calculate(BarSeries barSeries, TradingRecord tradingRecord) {
-        return tradingRecord.getPositions().stream()
+        return tradingRecord.getPositions()
+                .stream()
                 .filter(Position::isClosed)
                 .map(Position::getProfit)
                 .filter(Num::isPositive)
