+++ ta4j-core/src/test/java/org/ta4j/core/criteria/pnl/MaxPositionNetProfitCriterionTest.java
@@ -55,9 +55,7 @@ public class MaxPositionNetProfitCriterionTest extends AbstractCriterionTest {
 
     @Test
     public void calculateFindsLargestClosedProfit() {
-        var series = new MockBarSeriesBuilder().withNumFactory(numFactory)
-                .withData(1, 2, 3, 4, 5, 6)
-                .build();
+        var series = new MockBarSeriesBuilder().withNumFactory(numFactory).withData(1, 2, 3, 4, 5, 6).build();
         var amount = numFactory.one();
         var record = new BaseTradingRecord();
